third_party/blink/web_tests/html/selectmenu/selectmenu-typeahead-with-spacekey.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<selectmenu id="selectmenu" tabindex="0">
    <button type="button" slot="button" behavior="button"
             id="selectmenu-button">
        Button
    </button>
    <option>Canada</option>
    <option>Spain</option>
    <option>United Arab Emirates</option>
    <option>United States</option>
</selectmenu>
<script>
description('Two keystrokes are considered as part of one typehead session if time difference between them is less than 1 sec');

window.jsTestIsAsync = true;

function keyDown(key, modifiers)
{
    if (!window.eventSender)
      debug("No event sender");
    eventSender.keyDown(key, modifiers);
}

var popup = document.getElementById("selectmenu");
var button = document.getElementById("selectmenu-button");

function test1() {
    debug('1. space key as part of search string.');
    button.focus();
    popup.value = "Canada";
    keyDown('U');
    keyDown('n');
    keyDown('i');
    keyDown('t');
    keyDown('e');
    keyDown('d');
    keyDown(' ');
    keyDown('S');
    shouldBeFalse('popup.open');
    shouldBeEqualToString('popup.value', 'United States');
    button.blur();

    debug('2. space key as part of search string with some delay.');
    button.focus();
    popup.value = "Canada";
    keyDown('U');
    keyDown('n');
    keyDown('i');
    keyDown('t');
    keyDown('e');
    keyDown('d');
    keyDown(' ');
    internals.resetSelectMenuTypeAheadSession(popup);
    keyDown('S');
    shouldBeFalse('popup.open');
    shouldBeEqualToString('popup.value', 'Spain');
    button.blur();

    debug('3. space key to open popup menu.');
    button.focus();
    popup.value = "Canada";
    keyDown('U');
    keyDown('n');
    keyDown('i');
    keyDown('t');
    keyDown('e');
    keyDown('d');
    internals.resetSelectMenuTypeAheadSession(popup);
    keyDown(' ');
    shouldBeTrue('popup.open');
    shouldBeEqualToString('popup.value', 'United Arab Emirates');
    button.blur();

    finishJSTest();
}

test1();
</script>
</body>
